+++ modules/core/test/lib/api/set-loader-options.spec.js
@@ -1,7 +1,18 @@
 import test from 'tape-promise/tape';
-import {setLoaderOptions} from '@loaders.gl/core';
+import {setLoaderOptions, getLoaderOptions} from '@loaders.gl/core';
 
 test('setLoaderOptions', (t) => {
   setLoaderOptions({});
+  t.end();
+});
+
+test('getLoaderOptions', (t) => {
+  const options1 = getLoaderOptions();
+  t.notOk(options1.customOption);
+  setLoaderOptions({
+    customOption: 'customValue'
+  });
+  const options2 = getLoaderOptions();
+  t.equals(options2.customOption, 'customValue');
   t.end();
 });