+++ modules/textures/src/compressed-texture-writer.ts
@@ -1,6 +1,18 @@
-import type {Writer} from '@loaders.gl/loader-utils';
+import type {Writer, WriterOptions} from '@loaders.gl/loader-utils';
 import {VERSION} from './lib/utils/version';
 import {encodeImageURLToCompressedTextureURL} from './lib/encoders/encode-texture';
+
+export type CompressedTextureWriterOptions = WriterOptions & {
+  cwd?: string;
+  texture?: {
+    format: string;
+    compression: string;
+    quality: string;
+    mipmap: boolean;
+    flipY: boolean;
+    toolFlags: string;
+  };
+};
 
 /**
  * DDS Texture Container Exporter
@@ -28,4 +40,10 @@ export const CompressedTextureWriter = {
 };
 
 // TYPE TESTS - TODO find a better way than exporting junk
-export const _TypecheckCompressedTextureWriter: Writer = CompressedTextureWriter;
+export const _TypecheckCompressedTextureWriter: Writer & {
+  encodeURLtoURL: (
+    inputUrl: string,
+    outputUrl: string,
+    options?: CompressedTextureWriterOptions
+  ) => Promise<string>;
+} = CompressedTextureWriter;