+++ modules/textures/src/index.ts
@@ -1,3 +1,6 @@
+import {isBrowser} from '@loaders.gl/loader-utils';
+import {VERSION} from './lib/utils/version';
+
 // Types
 export type {GPUTextureFormat} from '@loaders.gl/schema';
 export type {TextureLoaderOptions} from './compressed-texture-loader';
@@ -10,7 +13,21 @@ export {NPYLoader, NPYWorkerLoader} from './npy-loader';
 
 // Writers
 export {CompressedTextureWriter} from './compressed-texture-writer';
-export {KTX2BasisUniversalTextureWriter} from './ktx2-basis-universal-texture-writer';
+export {KTX2BasisWriter} from './ktx2-basis-writer';
+
+export const KTX2BasisWriterWorker = {
+  name: 'Basis Universal Supercompressed GPU Texture',
+  id: isBrowser ? 'ktx2-basis-writer' : 'ktx2-basis-writer-nodejs',
+  module: 'textures',
+  version: VERSION,
+  extensions: ['ktx2'],
+  options: {
+    useSRGB: false,
+    qualityLevel: 10,
+    encodeUASTC: false,
+    mipmaps: false
+  }
+};
 
 // Texture Loading API
 export {loadImageTexture} from './lib/texture-api/load-image';